Move identity tests to their own test file
